apps/mcp/pyproject.toml
[tool.poetry]
name = "mcp-server"
version = "0.1.0"
description = "Custom MCP server with Arcade integration and custom tools"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "mcp_server", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
mcp = "^1.0.0"
arcadepy = "^1.0.0"
fastapi = "^0.115.0"
uvicorn = "^0.30.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
httpx = "^0.27.0"
python-dotenv = "^1.0.0"
structlog = "^23.2.0"
supabase = "^2.3.0"
click = "^8.1.0"
anyio = "^4.5.0"
starlette = "^0.39.0"
openai = "^1.91.0"
replicate = "^0.34.0"
google-cloud-storage = "==3.0.0"
google-crc32c = "==1.6.0"
requests = "^2.31.0"
sentry-sdk = {version = ">=2.8.0", extras = ["fastapi"]}
docling = "^2.51.0"
asyncpg = "^0.30.0"
pyyaml = "^6.0.2"
e2b-code-interpreter = "^1.0.0"
mem0ai = "^0.1.117"
aiohttp = "^3.10.0"
pyjwt = "^2.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.7.0"
pre-commit = "^3.5.0"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
mcp-server = "mcp_server.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 